anyways

+++ src/main/java/net/ccbluex/liquidbounce/injection/forge/mixins/entity/MixinEntityLivingBase.java
@@ -120,7 +120,7 @@ public abstract class MixinEntityLivingBase extends MixinEntity {
     }
 
     @ModifyConstant(method = "onLivingUpdate", constant = @Constant(doubleValue = 0.005D))
-    private double refactor1_9MovementThreshold(int constant) {
+    private double refactor1_9MovementThreshold(double constant) {
         if (ViaForge.getInstance().getVersion() != 47)
             return 0.003D;
         return 0.005D;
